move app-related classes to its own folder
