+++ lib/domain/apps/app_install_status.dart
